Fix for ES(1+ 1) and moved new implementation of jDE
