+++ NiaPy/algorithms/modified/__init__.py
@@ -1,8 +1,7 @@
 """Implementation of modified nature-inspired algorithms."""
 
 from NiaPy.algorithms.modified.hba import HybridBatAlgorithm
-from NiaPy.algorithms.modified.jde import SelfAdaptiveDifferentialEvolutionAlgorithm
-from NiaPy.algorithms.modified.dynnpjde import DynNPSelfAdaptiveDifferentialEvolutionAlgorithm
+from NiaPy.algorithms.modified.jde import SelfAdaptiveDifferentialEvolutionAlgorithm, DynNPSelfAdaptiveDifferentialEvolutionAlgorithm
 
 __all__ = [
     'HybridBatAlgorithm',
